+++ src/test/neural_network_test.rs
@@ -1,0 +1,17 @@
+use crate::neural_network::*;
+use approx::assert_relative_eq;
+use ndarray::prelude::*;
+
+mod activation_test;
+mod average_pooling_1d_test;
+mod average_pooling_2d_test;
+mod conv2d_test;
+mod flatten_test;
+mod global_average_pooling_2d_test;
+mod global_max_pooling_2d_test;
+mod layer_weight_test;
+mod loss_function_test;
+mod lstm_test;
+mod max_pooling_2d_test;
+mod optimizer_test;
+mod simple_rnn_test;