+++ src/test/utility_test.rs
@@ -1,0 +1,5 @@
+mod kernel_pca_test;
+mod linear_discriminant_analysis_test;
+mod principal_component_analysis_test;
+mod t_sne_test;
+mod train_test_split_test;